Removed duplicate pices of code in inference/io and clean up
